% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mstte.R
\name{rsimid}
\alias{rsimid}
\title{Simulate multi-state survival data}
\usage{
rsimid(dist01 = c("weibull", "exponential", "gompertz"),
  dist02 = c("weibull", "exponential", "gompertz"),
  dist12 = c("weibull", "exponential", "gompertz"), lambdas01, gammas01,
  betas01, lambdas02, gammas02, betas02, lambdas12, gammas12, betas12,
  hazard01, hazard02, hazard12, loghazard01, loghazard02, loghazard12, tde,
  tdefunction = NULL, mixture = FALSE, pmix = 0.5, hazard, loghazard,
  cumhazard, logcumhazard, x, cens, idvar = NULL, ids = NULL,
  nodes = 15, interval = c(1e-08, 500),
  seed = sample.int(.Machine$integer.max, 1), ...)
}
\description{
Simulate multi-state survival data
}
