% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_jmmstte.R
\name{sim_idm_jm}
\alias{sim_idm_jm}
\title{Simulate a illness-death set with longitudinal data}
\usage{
sim_idm_jm(n = 200, M = 1, n_trans = 3, fixed_trajectory = "cubic",
  random_trajectory = "linear", assoc = c("etavalue"),
  basehaz = c("weibull"), betaLong_intercept = 10,
  betaLong_binary = -1, betaLong_continuous = 1,
  betaLong_linear = -0.25, betaLong_quadratic = 0.03,
  betaLong_cubic = -0.0015, betaLong_aux = 0.5,
  betaEvent_intercept01 = -7.5, betaEvent_intercept02 = -6.5,
  betaEvent_intercept12 = -5.5, betaEvent_binary01 = -0.5,
  betaEvent_binary02 = -0.4, betaEvent_binary12 = -0.3,
  betaEvent_continuous01 = 0.5, betaEvent_continuous02 = 0.4,
  betaEvent_continuous12 = 0.3, betaEvent_assoc01 = 0.55,
  betaEvent_assoc02 = 0.57, betaEvent_assoc12 = 0.5,
  betaEvent_aux01 = 1.2, betaEvent_aux02 = 0.8,
  betaEvent_aux12 = 0.9, b_sd = c(1.5, 0.07), b_rho = -0.2,
  prob_Z1 = 0.5, mean_Z2 = 0, sd_Z2 = 1, max_yobs = 10,
  cens = c(199, 199), max_fuptime = 20, balanced = FALSE,
  family = gaussian, clust_control = NULL, return_eta = FALSE,
  seed = sample.int(.Machine$integer.max, 1), interval = c(1e-08, 200))
}
\description{
Simulate a illness-death set with longitudinal data
}
