% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dichotomize.dataset.R
\name{dichotomize.dataset}
\alias{dichotomize.dataset}
\title{Dichotomize a single dataset}
\usage{
dichotomize.dataset(x, split.at = NA)
}
\arguments{
\item{x}{A vector of values to be dichotomized}

\item{split.at}{An optional value that can be used to dichotomize instead of
median}
}
\value{
A vector of the data dichotomized onto a 0/1 (low/high) scale.
}
\description{
Split a dataset into two groups by median-dichotomization
}
\examples{

tmp <- data.frame(y = rnorm(100));
tmp$x <- dichotomize.dataset(tmp$y);

}
\author{
Syed Haider & Paul C. Boutros
}
\keyword{survival}
