% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.program.defaults.R
\name{get.program.defaults}
\alias{get.program.defaults}
\title{A utility function to return the inst/ directory of the installed package
and other default settings}
\usage{
get.program.defaults(networks.database = "default")
}
\arguments{
\item{networks.database}{Name of the pathway networks database. Default to
NCI PID/Reactome/Biocarta i-e "default"}
}
\value{
Returns a list of paths to the input directories/files where the
contents of this package are installed
}
\description{
A utility function to return the inst/ directory of the installed package to
get the test datasets and other program related data contents
}
\examples{

program.data <- get.program.defaults();

}
\author{
Syed Haider
}
\keyword{IO}
