% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.matrix.R
\name{make.matrix}
\alias{make.matrix}
\title{Utility function used by \code{get.adjacency.matrix()}}
\usage{
make.matrix(vertices, interactions)
}
\arguments{
\item{vertices}{Comma separated list of nodes}

\item{interactions}{Comma separated list of edges}
}
\value{
Returns adjacency matrix
}
\description{
Utility function used by \code{get.adjacency.matrix()}
}
\examples{

x1 <- make.matrix("a,b,c", "a:b,b:c");

}
\author{
Syed Haider
}
\keyword{Networks}
